Implement GET Weather data

+++ src/main/kotlin/weather/analyst/two/entity/repository/WeatherDataObjectRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository
 import weather.analyst.two.entity.WeatherDataObject
 
 @Repository
-interface WeatherDataObjectRepository : CrudRepository<WeatherDataObject, Int> {
+interface WeatherDataObjectRepository : CrudRepository<WeatherDataObject, Long> {
 }
